+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project, at least loosely, adheres to [Semantic Versioning](https://sem
 ## Unreleased
 ### Changed
 - index.txt is only checked at most once a day
+- Moved observatories to JSON file.  Changed way observatories are loaded/overloaded
 ### Added
 ### Fixed